requirements.txt
streamlit
pdfplumber
python-docx
pandas
openpyxl
faiss-cpu
sentence-transformers
openai
